backend/requirements.txt
flask==2.1.2
pymysql==1.0.2
datetime==4.5
python-dotenv==0.20.0
scikit-learn==1.0.2
attacut==1.0.6
joblib==1.1.0
numpy
tensorflow
gensim==4.1.2
pythainlp==3.0.8
pandas==1.2.4
flask_cors==3.0.10
requests==2.28.0
